+++ extensions/cran-e-search/CHANGELOG.md
@@ -1,3 +1,8 @@
 # CRAN/E Packages Changelog
 
-## [Initial Version] - 2023-02-22
+## [Feature: Github Trends] - 2023-02-23
+
+- New: Github Trends: repos by stars
+- New: Github Trends: authors by followers
+
+## [Initial Version] - 2023-02-20